Ignore docker-rm errors
When the docker-rm command fails due to a timeout the plugin throws an
exception. However, there are situations when these errors are common
and we don't know yet what their root cause is [0].

So, this patch simply ignores the error, letting the docker-rm command
fail without causing the build to fail.

Ref0: https://issues.jenkins-ci.org/browse/JENKINS-42322